-change type of coordinates -add method ResponseByGeo()

+++ Weather/Coordinates.cs
@@ -2,7 +2,7 @@ namespace WeatherBot
 {
     public class Coordinates
     {
-        public float lon { get; set; }
-        public float lat { get; set; }
+        public double lon { get; set; }
+        public double lat { get; set; }
     }
 }
